pyproject.toml
[project]
name = "bitter-lesson-cvpr"
version = "0.1.0"
description = "Some exploration on CVPR digest to see how community has embraced the \"Bitter Lesson\""
authors = [
    { name = "Mojtaba Yousefi", email = "mojtaba.yousefi@live.com" }
]
dependencies = [
    "beautifulsoup4>=4.12.3",
    "requests>=2.32.3",
    "selenium>=4.21.0",
    "semanticscholar>=0.8.2",
    "aiohttp>=3.9.5",
    "asyncio>=3.4.3",
    "tqdm>=4.66.4",
    "python-dotenv>=1.0.1",
    "magentic>=0.26.0",
    "webdriver-manager>=4.0.1",
    "ipykernel>=6.29.4",
    "plotly>=5.22.0",
    "pandas>=2.2.2",
    "nbformat>=5.10.4",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.rye.dependencies]
ipykernel = "*"
plotly = "*"
dash = "*"
pandas = "*"
openai = "*"
pydantic = "*"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/bitter_lesson_cvpr"]
